Beacons can return results

+++ server/core/core.go
@@ -24,8 +24,8 @@ import (
 )
 
 // EnvelopeID - Generate random ID of 8 bytes
-func EnvelopeID() uint64 {
+func EnvelopeID() int64 {
 	randBuf := make([]byte, 8) // 64 bits of randomness
 	rand.Read(randBuf)
-	return binary.LittleEndian.Uint64(randBuf)
+	return int64(binary.LittleEndian.Uint64(randBuf))
 }
